+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.0.5
+
+- fix: log display `name` of single element
+- fix: `el.r` save your original string when you use expression `${var}`
+
 ## 0.0.4
 
 - remove extending `cy` method by custom command